+++ relay-server/src/metrics_extraction/transactions/mod.rs
@@ -652,6 +652,7 @@ mod tests {
                 platform: ~,
                 was_transaction: ~,
                 kind: ~,
+                _performance_issues_spans: ~,
                 other: {},
             },
         ]